requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
click-shell==1.0
coverage==4.4.2
cryptography==2.0.2
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
idna==2.5
ipaddress==1.0.18
lxml==4.1.0
matplotlib==2.0.2
mock==2.0.0
msgpack-python==0.4.8
numpy==1.13.1
packaging==16.8
pbr==3.1.1
pycparser==2.18
pymetasploit==1.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-nmap==0.6.1
pytz==2017.2
pyvas==0.4.0
scapy==2.3.3
six==1.10.0
subprocess32==3.2.7
